Implementei o registro saida de itens, onde faz verificações para saber se tem a qtd disponivel,fazendo a logica no service, e implementação de ajuste no atualiza estoque

+++ BackEnd/QuataSafeGuardBackEnd/src/main/java/com/quata/quatasafeguardbackend/entities/RegistroSaidaItens.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class RegistroSaidaItens {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idRegistroSaidaItens;
+    private Long idRegistroSaidaItens;
     private Integer qtde;
 
     @ManyToOne
